+++ math/s_clog_template.c
@@ -32,7 +32,7 @@ M_DECL_FUNC (__clog) (CFLOAT x)
   if (__glibc_unlikely (rcls == FP_ZERO && icls == FP_ZERO))
     {
       /* Real and imaginary part are 0.0.  */
-      __imag__ result = signbit (__real__ x) ? (FLOAT) M_MLIT (M_PI) : 0;
+      __imag__ result = signbit (__real__ x) ? M_MLIT (M_PI) : 0;
       __imag__ result = M_COPYSIGN (__imag__ result, __imag__ x);
       /* Yes, the following line raises an exception.  */
       __real__ result = -1 / M_FABS (__real__ x);
@@ -94,7 +94,7 @@ M_DECL_FUNC (__clog) (CFLOAT x)
       else
 	{
 	  FLOAT d = M_HYPOT (absx, absy);
-	  __real__ result = M_LOG (d) - scale * (FLOAT) M_MLIT (M_LN2);
+	  __real__ result = M_LOG (d) - scale * M_MLIT (M_LN2);
 	}
 
       __imag__ result = M_ATAN2 (__imag__ x, __real__ x);
